requirements.txt
pip
watchdog
pyyaml
obs-websocket-py